Removed unnecessary dockerfiles_test.go. Moved tests out of cmd. Reduced testdata files
