+++ actuary/images_test.go
