+++ actuary/secops_test.go
