--- cmd/actuary/testdata/docker.service
@@ -1,2 +1,0 @@
-#!/bin/sh
-echo "docker.service"
