--- cmd/actuary/testdata/docker.socket
@@ -1,2 +1,0 @@
-#!/bin/sh
-echo "docker.socket"
