--- cmd/actuary/testdata/etc/docker/certs.d/certFolder/cert
@@ -1,2 +1,0 @@
-#!/bin/sh
-echo "cert"
