--- cmd/actuary/testdata/etc/docker/daemon.json
@@ -1,2 +1,0 @@
-#!/bin/sh
-echo "docker.socket"
