--- cmd/actuary/testdata/fstabFail
@@ -1,1 +1,0 @@
-/filler /wrong
