--- cmd/actuary/testdata/fstabPass
@@ -1,1 +1,0 @@
-/filler /var/lib/docker
