--- cmd/actuary/testdata/groupFail
@@ -1,1 +1,0 @@
-docker:users:
