--- cmd/actuary/testdata/groupPass
@@ -1,1 +1,0 @@
-docker:users: user1, user2, user3
