--- cmd/actuary/testdata/var/run/docker.sock
@@ -1,2 +1,0 @@
-#!/bin/sh
-echo "docker.sock"
