+++ oututils/oututils.go
@@ -4,7 +4,7 @@ import (
 	"encoding/json"
 	"encoding/xml"
 	"fmt"
-	"github.com/diogomonica/actuary/cmd/actuary"
+	"github.com/diogomonica/actuary/actuary"
 	"github.com/fatih/color"
 	"io/ioutil"
 	"log"
