+++ profileutils/profileutils.go
@@ -6,7 +6,6 @@ import (
 	"io/ioutil"
 	"log"
 	"net/http"
-
 	"github.com/BurntSushi/toml"
 )
